+++ src/App.js
@@ -1,17 +1,6 @@
 import React from 'react';
-
-
-const myList = [
-  {title: "Complete assignment",
-  objectID: 1,
-  },
-  {title: "Pospone assignment",
-  objectID: 2,
-  },
-  {title: "Assignment finished",
-  objectID: 3,
-  },
-];
+import TodoList from './TodoList'
+import AddTodoForm from './AddTodoForm';
 
 
 function App() {
@@ -21,15 +10,11 @@ function App() {
       <h1>Todo List</h1>
     </div>
     <hr />
-    <ul>
-        {myList.map(function (item) {
-          return <li key={item.objectID}>
-            <span>
-              {item.title}
-            </span>
-          </li>;
-        })}
-    </ul>
+
+    <AddTodoForm />
+    <TodoList />
+
+
     </>
   )
 }
