+++ src/TodoList.js
@@ -1,0 +1,29 @@
+import React from 'react'
+
+const myList = [
+    {title: "Complete assignment",
+    objectID: 1,
+    },
+    {title: "Pospone assignment",
+    objectID: 2,
+    },
+    {title: "Assignment finished",
+    objectID: 3,
+    },
+  ];
+
+export default function TodoList() {
+  return (
+    <div>
+        <ul>
+        {myList.map(function (item) {
+          return <li key={item.objectID}>
+            <span>
+              {item.title}
+            </span>
+          </li>;
+        })}
+    </ul>
+    </div>
+  )
+}
